feat: enhance SEO and metadata across the application
- Updated metadata titles and descriptions for improved SEO, including specific adjustments for the blog, colors, and pro pages.
- Added support for additional user agents in the robots.txt to optimize crawling for AI and answer-engine bots.
- Introduced new sitemap entries for llms.txt and agents.md to enhance discoverability.
- Refactored the installation page to include a SEOWrapper for better SEO management.
- Updated package.json to include @vercel/functions for improved serverless function handling.
