+++ package.json
@@ -67,6 +67,7 @@
     "@types/stripe": "^8.0.416",
     "@upstash/redis": "^1.34.3",
     "@vercel/analytics": "^1.3.1",
+    "@vercel/functions": "^3.7.5",
     "@vercel/speed-insights": "^1.0.14",
     "axios": "^1.7.9",
     "canvas-confetti": "^1.9.3",
